Initial points prototyping commit.
The new points/ directory contains all the prototyping code for getting large
numbers of points onto a map. We're using an App Engine backend to with a new
points API that pushes pages of points to the JS client through the App Engine
Channel API.

This commit includes a functional backend with a stubbed in points API that
pushes pages of mock points back to the client. Next step is to use the
App Engine Async URL Fetch API to grab pages of points in parallel from
a local instance of the VertNet app preloaded with thousands of points. As
each page is received by the backend, it will get pushed through the Channel
API back to the client where it will get rendered on the map.

At some point it will be too many points, and then we'll roll in the points
canvas code. At some point that will fail too, and then we'll get into
CouchDB bounding box queries and tiling.